extract java version code to own class
